+++ vspec/model/vsstree.py
@@ -12,39 +12,11 @@ import sys
 import re
 import copy
 from anytree import Node, Resolver, ChildResolverError
+from .exceptions import IncompleteElementException, NameStyleValidationException, ImpossibleMergeException, UnknownAttributeException
 
-from .constants import VSSType, VSSDataType, StringStyle, Unit
+from .constants import VSSType, VSSDataType, Unit
 
 DEFAULT_SEPARATOR = "."
-
-
-class UnknownAttributeException(Exception):
-    def __init__(self, message):
-
-        # Call the base class constructor with the parameters it needs
-        super().__init__(message)
-
-
-class NameStyleValidationException(Exception):
-    def __init__(self, message):
-
-        # Call the base class constructor with the parameters it needs
-        super().__init__(message)
-
-
-class ImpossibleMergeException(Exception):
-    def __init__(self, message):
-
-        # Call the base class constructor with the parameters it needs
-        super().__init__(message)
-
-
-class IncompleteElementException(Exception):
-    def __init__(self, message):
-
-        # Call the base class constructor with the parameters it needs
-        super().__init__(message)
-
 
 class VSSNode(Node):
     """Representation of an VSS element according to the vehicle signal specification."""
@@ -287,16 +259,3 @@ class VSSNode(Node):
                 raise UnknownAttributeException(
                     "Invalid VSS element %s, only attributes can use default" % name)
 
-
-def camel_case(st):
-    """Camel case string conversion"""
-    s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', st)
-    s2 = re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower()
-    return re.sub(r'(?:^|_)([a-z])', lambda x: x.group(1).upper(), s2)
-
-
-def camel_back(st):
-    """Camel back string conversion"""
-    s1 = re.sub('(.)([A-Z][a-z]+)', r'\1_\2', st)
-    s2 = re.sub('([a-z0-9])([A-Z])', r'\1_\2', s1).lower()
-    return re.sub(r'_([a-z])', lambda x: x.group(1).upper(), s2)
